+++ src/SDK/Metrics/MeterProviderBuilder.php
@@ -5,10 +5,8 @@ declare(strict_types=1);
 namespace OpenTelemetry\SDK\Metrics;
 
 use OpenTelemetry\SDK\Common\Attribute\Attributes;
-use OpenTelemetry\SDK\Common\Future\CancellationInterface;
 use OpenTelemetry\SDK\Common\Instrumentation\InstrumentationScopeFactory;
 use OpenTelemetry\SDK\Common\Time\ClockFactory;
-use OpenTelemetry\SDK\Common\Util\ShutdownHandler;
 use OpenTelemetry\SDK\Metrics\Exemplar\ExemplarFilter\WithSampledTraceExemplarFilter;
 use OpenTelemetry\SDK\Metrics\StalenessHandler\ImmediateStalenessHandlerFactory;
 use OpenTelemetry\SDK\Metrics\View\CriteriaViewRegistry;
@@ -20,7 +18,6 @@ class MeterProviderBuilder
     // @var array<MetricReaderInterface>
     private array $metricReaders = [];
     private ?ResourceInfo $resource = null;
-    private bool $autoShutdown = false;
 
     public function registerMetricReader(MetricReaderInterface $reader): self
     {
@@ -44,21 +41,11 @@ class MeterProviderBuilder
     }
 
     /**
-     * Automatically shut down the tracer provider on process completion. If not set, the user is responsible for calling `shutdown`.
-     */
-    public function setAutoShutdown(bool $shutdown): self
-    {
-        $this->autoShutdown = $shutdown;
-
-        return $this;
-    }
-
-    /**
      * @psalm-suppress PossiblyInvalidArgument
      */
     public function build(): MeterProviderInterface
     {
-        $meterProvider = new MeterProvider(
+        return new MeterProvider(
             null,
             $this->resource ?? ResourceInfoFactory::emptyResource(),
             ClockFactory::getDefault(),
@@ -69,10 +56,5 @@ class MeterProviderBuilder
             new WithSampledTraceExemplarFilter(),
             new ImmediateStalenessHandlerFactory(),
         );
-        if ($this->autoShutdown) {
-            ShutdownHandler::register(fn (?CancellationInterface $cancellation = null): bool => $meterProvider->shutdown());
-        }
-
-        return $meterProvider;
     }
 }
