+++ src/SDK/Trace/TracerProviderBuilder.php
@@ -4,12 +4,7 @@ declare(strict_types=1);
 
 namespace OpenTelemetry\SDK\Trace;
 
-use OpenTelemetry\SDK\Common\Future\CancellationInterface;
-use OpenTelemetry\SDK\Common\Util\ShutdownHandler;
 use OpenTelemetry\SDK\Resource\ResourceInfo;
-use OpenTelemetry\SDK\Resource\ResourceInfoFactory;
-use OpenTelemetry\SDK\Trace\Sampler\AlwaysOnSampler;
-use OpenTelemetry\SDK\Trace\Sampler\ParentBased;
 
 class TracerProviderBuilder
 {
@@ -17,7 +12,6 @@ class TracerProviderBuilder
     private ?array $spanProcessors = [];
     private ?ResourceInfo $resource = null;
     private ?SamplerInterface $sampler = null;
-    private bool $autoShutdown = false;
 
     public function addSpanProcessor(SpanProcessorInterface $spanProcessor): self
     {
@@ -33,16 +27,6 @@ class TracerProviderBuilder
         return $this;
     }
 
-    /**
-     * Automatically shut down the tracer provider on process completion. If not set, the user is responsible for calling `shutdown`.
-     */
-    public function setAutoShutdown(bool $shutdown): self
-    {
-        $this->autoShutdown = $shutdown;
-
-        return $this;
-    }
-
     public function setSampler(SamplerInterface $sampler): self
     {
         $this->sampler = $sampler;
@@ -52,15 +36,10 @@ class TracerProviderBuilder
 
     public function build(): TracerProviderInterface
     {
-        $tracerProvider = new TracerProvider(
+        return new TracerProvider(
             $this->spanProcessors,
-            $this->sampler ?? new ParentBased(new AlwaysOnSampler()),
-            $this->resource ?? ResourceInfoFactory::defaultResource(),
+            $this->sampler,
+            $this->resource,
         );
-        if ($this->autoShutdown) {
-            ShutdownHandler::register(fn (?CancellationInterface $cancellation = null): bool => $tracerProvider->shutdown($cancellation));
-        }
-
-        return $tracerProvider;
     }
 }
